+++ src/main/resources/META-INF/services/org.apache.spark.sql.sources.DataSourceRegister
@@ -1,1 +1,1 @@
-com.databricks.spark.xml.DefaultSource
+com.darkrows.spark.xml.DefaultSource
