+++ src/main/scala/com/darkrows/spark/xml/table/XmlColumn.scala
@@ -1,0 +1,6 @@
+package com.darkrows.spark.xml.table
+
+import net.sf.saxon.s9api.XPathExecutable
+
+
+case class XmlColumn(name: String, xpath: String)
