+++ src/main/scala/com/darkrows/spark/xml/table/XmlNamespace.scala
@@ -1,0 +1,3 @@
+package com.darkrows.spark.xml.table
+
+case class XmlNamespace(prefix: String, uri: String)
