--- src/main/scala/com/databricks/spark/xml/table/XmlColumn.scala
@@ -1,6 +1,0 @@
-package com.databricks.spark.xml.table
-
-import net.sf.saxon.s9api.XPathExecutable
-
-
-case class XmlColumn(name: String, xpath: String)
