--- src/main/scala/com/databricks/spark/xml/table/XmlCompiledColumn.scala
@@ -1,9 +1,0 @@
-package com.databricks.spark.xml.table
-
-import net.sf.saxon.s9api.XPathExecutable
-import org.apache.spark.sql.types.DataType
-
-case class XmlCompiledColumn(name: String, xpath: XPathExecutable, dataType: DataType)
-
-
-
