--- src/main/scala/com/databricks/spark/xml/table/XmlNamespace.scala
@@ -1,3 +1,0 @@
-package com.databricks.spark.xml.table
-
-case class XmlNamespace(prefix: String, uri: String)
